+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/experiments/VariantHandler.kt
@@ -1,0 +1,6 @@
+package com.mercadopago.android.px.internal.experiments
+
+interface VariantHandler {
+    fun visit(variant: PulseVariant)
+    fun visit(variant: BadgeVariant)
+}