+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/viewmodel/CreditCardDescriptorModel.java
@@ -10,6 +10,7 @@ import android.widget.TextView;
 import com.mercadopago.android.px.R;
 import com.mercadopago.android.px.internal.util.RateType;
 import com.mercadopago.android.px.internal.util.TextUtil;
+import com.mercadopago.android.px.internal.util.ViewUtils;
 import com.mercadopago.android.px.internal.util.textformatter.AmountLabeledFormatter;
 import com.mercadopago.android.px.internal.util.textformatter.CFTFormatter;
 import com.mercadopago.android.px.internal.util.textformatter.PayerCostFormatter;
@@ -75,6 +76,14 @@ public final class CreditCardDescriptorModel extends PaymentMethodDescriptorView
         @NonNull final TextView textView) {
         super.updateRightSpannable(spannableStringBuilder, textView);
         updateInstallmentsInfo(spannableStringBuilder, textView.getContext());
+    }
+
+    @Override
+    public void updateDrawableBackground(@NonNull final TextView textView) {
+        super.updateDrawableBackground(textView);
+        if (payerCostSelected == PayerCost.NO_SELECTED && installmentsRightHeader != null) {
+            ViewUtils.setDrawableBackgroundColor(textView, installmentsRightHeader.getBackgroundColor());
+        }
     }
 
     private void updateInstallmentsInfo(@NonNull final SpannableStringBuilder spannableStringBuilder,