+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 _12_06_2020_
 * ENHANCEMENT - Update CFTFormatter for CFTNA
 
+## VERSION 4.48.1
+_12_06_2020_
+* ENHANCEMENT - Update CFTFormatter for CFTNA
+
+## VERSION 4.46.4
+_12_06_2020_
+* ENHANCEMENT - Update CFTFormatter for CFTNA
+
+## VERSION 4.48.0
+_11_06_2020_
+* FEATURE - experiments with installments row
+* ENHANCEMENT - persist payment configuration
+* FIX - top fragment position in default congrats order
+
 ## VERSION 4.47.0
 _04_06_2020_
 * FEATURE - Send security type of payment as header of payment wrapper api call